requirements.txt
streamlit
torchvision
torch
pillow
